Remove special handling of negative numbers
When a negative number was given on the command line, it would have been parsed
as an argument, although it starts with a dash.

This caused surprising behaviour where an error would be given when grouping
short options, starting with a digit:

```
$ cmd -1 -2 -3
OK
$ cmd -123
ERROR: unknown argument -123
$ cmd -A1
OK
$ cmd -1A
ERROR: unknown argument -1A
```

Using the `--` to explicitly stop option parsing also fixes this problem, in general.

See #239.